refactor: move process observability signals (active_calls, loop_lag) out of pipecat into services/observability
