+++ api/app.py
@@ -70,7 +70,7 @@ async def lifespan(app: FastAPI):
 
         # Event-loop lag gauge — per-pod saturation signal read off
         # /health/active-calls during autoscaling load tests (scripts/loadtest).
-        from api.services.pipecat import loop_lag
+        from api.services.observability import loop_lag
 
         loop_lag.start()
